enable progress bar on all UNIX platforms
The progress bar rendered at the base of ongoing output from Cargo is
presently only drawn on a subset of UNIX platforms: FreeBSD, Linux, and
Mac OS.  The functionality required is basic, and essentially
universally available: "ioctl()", "TIOCGWINSZ", and "STDERR_FILENO".

When your platform is not in this list, the difference in behaviour is
subtle and the exact mechanism is difficult to locate.  It would be
better to fail to build on new ports until these definitions become
available in the libc crate for those platforms.