Slightly rewrite how conflicting activation errors are processed.
I'm not sure why the original code partitioned the errors in the way
that it did. I think it is better to exhaustively match all the reasons,
so that when new reasons are added, it will be clear that this code
needs to be updated. I also think it simplifies the code a little.